Empiezo operaciones add, div, sub, mul, store

+++ p02/copia/estrategia_almacenamiento/estrategia_almacenamiento.h
@@ -5,7 +5,7 @@
 // Interfaz para la estrategia de almacenamiento
 class EstrategiaAlmacenamiento {
  public:
-  virtual void almacenarDato(std::vector<std::vector<int>>&, const int, const int, const int) const = 0;
-  virtual const int obtenerDato(const std::vector<std::vector<int>>&, const int, const int) const = 0;
+  virtual void almacenarDato(std::vector<std::vector<long double>>&, const long double, const int, const int) const = 0;
+  virtual const long double obtenerDato(const std::vector<std::vector<long double>>&, const int, const int) const = 0;
   virtual ~EstrategiaAlmacenamiento() {}
 };
