+++ p02/copia/estrategia_almacenamiento/estrategia_almacenamiento_unico.h
@@ -5,12 +5,12 @@
 // Estrategia actual: Cada registro almacena un único espacio para datos
 class EstrategiaUnicoEspacio : public EstrategiaAlmacenamiento {
  public:
-  void almacenarDato(std::vector<std::vector<int>>& registros, const int kDato,
+  void almacenarDato(std::vector<std::vector<long double>>& registros, const long double kDato,
                      const int kIndiceFila, const int kIndiceColumna = 0) const override {
     registros[kIndiceFila][0] = kDato;
   }
-  const int obtenerDato(const std::vector<std::vector<int>>& kRegistros, const int kIndiceFila,
-                  const int kIndiceColumna = 0) const override {
+  const long double obtenerDato(const std::vector<std::vector<long double>>& kRegistros, const int kIndiceFila,
+                                const int kIndiceColumna = 0) const override {
     return kRegistros[kIndiceFila][0];
   }
 };
